minor improvemets

+++ src/Controller/TrainTripController.php
@@ -22,6 +22,7 @@ class TrainTripController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CHAIN_MANAGER');
         $trainTrip = new TrainTrip();
         $form = $this->createForm(TrainTripType::class, $trainTrip);
         $form->handleRequest($request);
@@ -42,6 +43,7 @@ class TrainTripController extends AbstractController
      */
     public function index(TrainTripRepository $truckRouteRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CHAIN_MANAGER');
         return $this->render('train_trip/index.html.twig', [
             'train_trips' => $truckRouteRepository->getAll(),
         ]);
@@ -52,6 +54,7 @@ class TrainTripController extends AbstractController
      */
     public function edit(Request $request, TrainTrip $truckRoute): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CHAIN_MANAGER');
         $form = $this->createForm(TrainTripType::class, $truckRoute);
         $form->handleRequest($request);
 
@@ -72,6 +75,7 @@ class TrainTripController extends AbstractController
      */
     public function delete(Request $request, TrainTrip $truckRoute): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CHAIN_MANAGER');
         if ($this->isCsrfTokenValid('train_trip', $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($truckRoute);
